mapped NumPy 'i' and 'I' types to intc and uintc
In the dictionary mapping letter codes I had to 
hardcode 'i' to int32 and 'I' to uint32. However,
it is semantically more appropriate to use "C int"
and "C unsigned int", as that's the meaning in
NumPy. As intc and uintc are now available, it
makes sense to update the table.