make typeinfer fail when a given loop is not supported
using the new ufunc_db, it uses information about supported loops
to raise a TypingError if a given loop is not supported. This is done
only for ufuncs relying in the ufunc_db for lowering, for the rest, a
type based mechanism is used (the one used before). The ufunc_db method
allows for a more fine grained specification.