Allow the dispatcher to keep references to types
When the dispatcher uses typecode_fallback to look up the typecode
of a type, it is possible for the object with the typecode it
returns to be GC'd before typecode_fallback is finished, therefore
returning a typecode for a now non-existent type.

This generally can occur in the first execution of a jitted
function, and is not problematic, except when typecodes are to be
cached - if the typecode is stored for future use, it cannot
legitimately be used to match argument types in future.

This commit makes it possible to prevent deletion of a type used
by typecode_fallback, which will be required to correct the
behaviour of the typecode cache for record types. When
typecode_fallback is called with retain_reference=1, it avoids
calling Py_DECREF on the type it obtains. The reference to the
type is not retained as it is not used from the dispatcher code -
it only needs to prevent deletion.