Fix caching of record arrays and add for scalars
The implementation of caching for ndarrays of structured types
erroneously uses the type_num to distinguish between different
types - unfortunately in the case of structured arrays, the
type_num is always NPY_VOID, so it treats all structured array
types as being the same type.

As a more correct alternative, this commit changes the
implementation to compare structured types based on the structure
of elements. It considers the name, type, and offset of fields in
a record, and if these match for all fields of two types then they
are considered equal.

Additionally, this commit adds caching for scalar record types
using similar logic.