Shorten names of records and structured arrays
Mangling the str of structured types can cause the names of
functions to become extremely long (long enough to crash NVVM with
a segfault).

This commit modifies the mangling to use the typecodes of
structured types rather than a full description of all their
fields.

The test attempts to ensure correctness by checking that no field
names form part of the transformed name, and that the transformed
name is not excessively long for any other reason.