Issue #504: Add JIT support for LIST_APPEND.
This patch adds support for JIT'ing LIST_APPEND, which will enable the
use of list comprehensions with Numba*.  LIST_APPEND is unique in that it
requires peeking back in the stack to grab the list that is appended to.
Dataflow analysis was extended to accomodate this.  Additionally, the
controlflow modeling of FOR_ITER had to be tweaked to represent the fact
that items are popped on the iterator exhaustion edge.

* Note that list comprehensions are still not supported for Python 3 as
  MAKE_FUNCTION is not supported yet.