Add support for sinc() function.
This adds support for the cardinal sine, sinc(), function.
The method in numpy is of type 'function' opposed to 'ufunc' hence
the implementation within is merely ufunc like.

Tests aim to cover the numerical domain and to trigger branches to
match outputs of numpy as the input approaches numerical 0. The
standard ufunc harness is used to check computations in when in
array context.