Add abs tol support to comparison. Fix `sinc()`.
`sinc()` was failing on a test machine via a bizarre numerical problem, it was
decided that supporting comparison by absolute tolerance may be useful
especially for values near zero, this patch introduces this feature. It also
fixes `sinc()` tests as a result. `sinc()` was failing for the value 1.e0 which
lead to `sin(pi)/pi` being evaluated, which naturally, due to numerical methods,
does not come out as 0.e0. The specific result value is below machine precision
and as demonstrated by failing code, somewhat machine/compiler/libm specific.