Add numpy.linalg.qr
This adds support for `numpy.linalg.qr`, the QR decomposition.

Only the first argument is supported as the kwarg is a string,
which is not supported yet. This means `qr()` follows the default
behaviour of operating in the `'reduced'` mode (see notes in
numpy, the behavior changed in 1.8 leading to `"reduced"` and
`"full"` aliasing).

The testing compares the implementation to numpy and falls back to
testing via reconstruction.