Correct type bug and provide slots for workspaces.
This fixes a couple of cases where a `double *` was used when a
`void *` ought to have been. It also fixes a recent failure case
with MKL 11.3.3 where the provision of a `rwork` or `iwork`
pointing to `NULL` is considered an error (`info` arg reports a
negative number corresponding to their argument position on
calling). The temporary fix is simply to provide a slot to which
they can point. Routines effected are: `geev` (eigenvalue/vector
driver) and `gesdd` (svd D&C driver).