Potential fix for "cannot compute fingerprint of empty X" #2152
This patch is one way of seemingly fixing the inability to compute
the fingerprint of an empty list (or set!), as reported in #2152.

It seems like empty containers can take part in fast lookup but
only if more op codes are added. Whether this is a valid approach,
given possible nuances, and the general "adding more opcodes"
going against the principle of fast lookup, is to be determined.

It should be noted that using '[' for an empty list seemingly
works too.