Ensure consistent iteration order of blocks for type inference.
When undertaking type inference the order in which the blocks were
walked was based on the iteration of a dictionary, this meant that
constraint network ordering was subject to the dictionary iteration
order. As a result, the same function IR could result in different
inferred types across Python versions at the whim of dictionary
ordering. This patch fixes this by storing the IR blocks in an
OrderedDict ordered ascending by the IR label.