Implement np.take and ndarray.take
This implements np.take and ndarray.take for basic integer based
indexing using ndarrays as the index, it also supports scalar and
single indexed lists as indexes.

There is a slight inefficiency in having to copy the index array
if it is F ordered to ensure C ordering for walking with nditer,
this is due to nditer not supporting flags at present.

Closes #2466.