Adding support for neighborhood stencil option.  If present, do not attempt to infer the kernel size but use the specified neighborhood option instead.
Fix typo for single-dimensional stencils.

Change kernel inference to compute min and max in all dimensions instead of
just one overall absolute min/max.  Change the generated code to use this
computed neighborhood (or the one passe din if the neighborhood option is
specified) to set the loop range on each dimension.