This commit fixes issues with the return type of stencils both in native mode and parallel=True mode.  We have to infer the return type of the specified kernel and make that the return type of the generated stencil function.  Switching to use np.full to generate new arrays in the stencil functions as all the numpy function *_like return arrays of the same type as the input even if dtype is otherwise specified.
Add function to calculate the return type of a stencil kernel by getting
the IR for the kernel and running it through type inference.  The negative
absolute indices in the stencil kernel are fine because they are interpreted
as indices from the back of the array.

Use that return to set the return type of the stencil function both in
native mode and parallel=True.

Fix for when the output array is specified and we have to get its type
and make that the return type of stencil.

Pass in the StencilFunc IR node to get_stencil_blocks.

mk_stencil_parfor takes return type now as well.