More locks...
A number of tests are still intermittently failing due to concurrent
access made to the LLVM context and LLVM in general. The three most
common offenders have already been patched, this attempts to fix most
of the remaining!

To work out where to put locks, llvmlite was instrumented with a
special decorator applied to all functions that make FFI calls to LLVM.
This decorator was context managed such that it would walk the stack
prior to yield and check that there was a llvmthreadsafe.py lock
present and raise if not. Running the test suite with this enabled
highlighted where locks needed to go. Since the llvmthreadsafe locks are
reentrant there's no harm in them being nested which reduced complexity
in applying them.

To help with tracing the stacks, this patch moves the functions that
had locks and called the LLVM API via llvmlite in
numba/llvmthreadsafe.py to numba/llvmthreadsafeapi.py so that the
presentation of a lock is consistent for all stacks (i.e. it's a
tracable import with consistent location not a local).

Once the locks were in place a couple of other issues materialised, due
to the difference in test run times caused by all the stack walking,
some tests relying on on disk cache failed when run in parallel mode due
to multiple threads writing/reading with a much longer interval between
the operations, as such reliance on e.g. testing properties of a single
cache location failed.

Limitations: The safety of LLVM calls at interpreter shutdown and during
more involved GC events is not necessarily covered as walking the
stack becomes more of a struggle during these times. Further, garbage
collection/deletion of the module level LLVM calls in
llvmthreadsafeapi.py trigger calls to an explicit `__del__` on the
execution engine wrapper which hits the LLVM FFI layer, these calls do
not have a lock guarding them at present as no simple hooks exist in
Numba for performing a guarded delete during e.g. interpreter shutdown.

This patch may or may not be applied. Given the limitations noted it may
make more sense to defer these fixes and patch llvmlite instead and
concentrate on fixing Python side synchronisation issues inside Numba.