Save typemap and calltypes in get_return_type and cache them so that copy propagation can be done later.
More advanced calculation of allowable constant tuple indices that use a build_tuple.
Now support copy propation in a kernel of the form: "t = 1; return a[t,0]".

For parfors, check if an index is a Var before iterating across it and just return
a failure if the neighborhood needs to be calculated and the index is just some tuple
var.