Created a function raise_if_incompatible_array_sizes that checks if all relatively indexed arrays are the same number of dimensions and at least as large as the first input array.
Discover all the relatively indexed arrays.
If there are more than 1 then call the above function at the start of the kernel to
raise an exception if the extra arrays are not appropriately sized.

Raise exception if there are no relatively indexed arrays.

Raise exception if the first argument is not a relatively indexed array.

Raise exception if an array name in the standard indexing decorator option is not a
parameter to the stencil kernel.