Fixes for stencil tests and cmath sin().
test_stencil has a declaration at the module level which causes
test discovery to fail on unsupported platforms and as a result
breaks incremental conda build recipes, included is a fix.

cmath sin{,h}() is sometimes generated via `sincos` intrinsic in
LLVM 5.x which has a slightly different value near `pi`, it is
near zero and near machine precision, this adds an absolute
tolerance to the tests to accommodate.