Add rtsys init guard and test.
This adds a guard to the methods on the rtsys instance to ensure that
the rtsys is initialized prior to use (lack of init leads to null ptr
access... segfault).

The failing-in-isolation tests reported in #2848 now have an indirect
call to `rtsys.initialize()` made by virtue of invoking
`numba.targets.cpu.CPUContext` in the test class `setUp`, this
initializing the rtsys correctly.

Fixes #2848