Add annotation info to dispatcher object.
This exposes the annotation information via a method on the
dispatcher. It also removes markup specific information from the
annotation dictionary and delegates the responsibility of
performing markup to the consumer. As a result the HTML annotation
path is now split from that which generates the annotation
dictionary which means annotations are easier to consume and don't
involve files.