Attempt to fix #2386.
This attempts to fix #2386 by adding code to the analysis of
loop lifting candidates that removes variables that are not used
in a lifted loop from the input/output to the function containing
the loop. This means more things get lifted in native mode as
spurious dependencies on objects are removed.

Fixes #2386