Add method of switching off colored error messages.
This patch sets up the test runners to run in "developer mode" and
also switches off color error messages in the event of failure.
This is so the most plain errors appear in logs with full
backtraces.

It also fixes up an issue in warning handling as NumPy ABI
difference warnings were causing a test to fail.