Pairwise sum implementation.
This switches the core array.sum() routine to behave more like
NumPy's, i.e. chunked and pairwise sum, for improved accuracy.

The chunk size is a magic number, 128 seems to be good locally.
A tree reduction is used to reduce the accumulated chunk sums, this
may or may not be a good idea (a repeated fold-and-add might be
faster and will use less memory).