Compute np.var and np.std correctly for complex types.
This fixes the method for computing np.var() and by induction
np.std() to correctly accomodate complex input values. It also
augments the test suite for reductions to include complex values
as inputs where valid. Further the same fix is made for parfors
and the tests are updated there to include complex types too.

Fixes #2692