Fixes raising of user defined exceptions for exec(<string>).
This fixes the handling of user defined exceptions in the case where
the source code is missing and therefore location information cannot
be obtained via the linecache. In practice this may be from e.g.
`exec(<string>)`, the REPL or modules with stripped source. In an
attempt to both match the CPython interpreter and provide users with
as much information as possible, the function name if available is
used in the location information, along with e.g. "<string>" as the
source and a known line number. If this information is not available
the fallback is to use a simple message declaring that no source is
available.

Fixes #3428