Handle User Defined Exceptions with args.
This updates the exception handler to correctly handle cases
where a user has defined an exception which implements a custom
`__init__` that may:

* Call super() with all its arguments.
* Call super() with none (or potentially some) of its
  arguments.

This also adds tests to check that the user defined exceptions raised
are of the correct type, and are also equal in value.

In addition the C code is updated to inject a location based frame into
exceptions that arise whilst attempting to handle exceptions (e.g. user
gave wrong arguments to a user defined exception at the call site). This
makes it such that even mistakes in generating exceptions get reported
from npm.

Fixes #3613