reword and cleanup
As discussed, the motivation of this guide was unclear, because it
wasn't stated clearly at the beginning what the problem statement is and
why `overload` is needed over  just plain `jit`. The story now goes as
follows: `set_to_x` is a library function that two people are using
(differently) in their algorithms. Hence we can treat `set_to_x` as a
library routine that needs an `overload`. Hopefully this will improve
coherence of the narative for the reader.