better error messages for np.transpose and tuples
Fixes #4053

NumPy 'transpose' does accept tuples but Numba doesn't and there is a
good reason for this. The problem is, that we don't know the shape of
the arrays contained within the tuple during the type inference phase.
If they have inhomogeneous shapes, we may end up with a ragged/jagged
array. We could check all tuples and their shapes at runtime but that
will be cumbersome. Instead, we issue a better warning instead for now.

For those wanting to use `np.argwhere` using the following snippet:

```
np.transpose(np.nonzero(a))
```

The appropriate workaround is:

```
np.transpose(np.vstack(np.nonzero(a)))
```