initial refcounting implementation
This includes the changes needed at the C and compiler level, including
all relevant function calls such as increment, decrement and setting up
the methods table. Also includes the equality method, but that isn't yet
used because neither count, contains or any of the other methods are
implemented just yet.