fixed nasty bug in pop
The memory copying wasn't taking into account the itemsize. But the
tests didn't catch this, because they were only testing for an itemsize
of one, in which case the code was correct. This was noticed during the
implementation for remove. This fixes the bug and adds a low-level test
for pop that uses an itemsize of 4.