implement using slice in getitem
It turns out, that handling a slice correctly isn't trivial. This commit
contains all the code to handle slicing via getitem (hopefully)
correctly. A large amount of unit tests is included due to the subtle
nature of slices. Only changes to the compiler part were needed and the
interprter part worked without modification.