fix checking status of malloc/realloc
As it turns out, both malloc and realloc on Linux may return either NULL or a
pointer that can be handed to free() if size is zero, at least according to
this source:

https://linux.die.net/man/3/realloc

This means, we make all allocations checks explicit and improve the check, by
additionally checking, that the size was non-zero.

When freeing the list, we also check that items is non-NULL before attempting
to free it.