accelerate setitem with slice
When using setitem with a slice and an item that is smaller than the
slice, we need to remove items. Until now, this was done with pop, which
actually incurred a memory move of the tail of the list for every item
being removed. Now that we can mass delete items in the list with `del`
and slices, we used that functionality instead.