implement list.index
This almost replicates cpython behaviour. The big difference is, that
the defaults for the start and end arguments are None rather than 0 and
MAX_INT64 (or maybe MAX_INT32?!). I wasn't sure where to obtain these
values in a platform independent way, but this could probably be changed
quite easily.