fix extending self
The code `l.extend(l)` wasn't working properly. What was needed, is a
defensive copy in case you try to extend the list with itself. The
problem with that was, that the `is` operator wasn't implemented for the
typed-list yet. So, this commit fixes the `is` operator for the typed
list and uses that to implement the defensive copy in case of self
referential extension. Tests included.