implement deeper compare
This implements the operators <, <=, > and >= for the typed list. Lists
with nested Numpy arrays can no longer be compared. This doesn't work in
cpython also, so it isn't too bad. The error message is a bit cryptic
however as this fails during lowering. Also, we now raise a TypingError
if the thing being compared to is not also a typed-list. The problem
here is that if we always just return False for a type-mismatch, then
both a < b and a > b will be False, which doesn't make any sense.  This
might be a bit over-restrictive however, but should be fine for most
use-cases.