Prevent CUDA kernel launch without a specified launch config.
This patch prevents the launch of a CUDA kernel with no
configuration as this causes confusion for (especially) first time
users. Current behaviour is that if no launch config is specified
then a default everything-set-to-1 config is used, new behaviour is
that if no launch config is specified then an exception is raised
pointing users to the syntax and documentation.