fix and test for using default args
When calling a function like:

```
@njit
def func(a, b=None):
    # code
```

We also need to call it like:

```
func(1)
```

To make sure that we don't reject a plain None when b is a NoneType.

In this case, this test was missing, but could be fixed by using
`is_nonlike` to perform typechecking.