Fix Python 2.7 Build on MacOS on Azure
At some point in October 2019 the Python 2.7 Builds on MacOSX started
failing with the following error:

```
/Users/runner/miniconda3/envs/travisci/include/python2.7/Python.h:33:10: fatal error: 'stdio.h' file not found
         ^~~~~~~~~
1 error generated.
```

While it isn't clear where the root cause of this issue lies --- either
with the Anaconda based c-compilers  or with the Microsoft build image
--- this patch

It is to be suspected that the following backport may be involved:

https://github.com/python/cpython/pull/12349

This also means that future releases of Python may actually suffer from
the same issue, so we will have to keep an eye on these.

Hopfully, this patch will no longer be required when Python 2.7 is EOL
in 2020.