Attempt at getting parfor to ignore boundscheck for slices
Parfor converts slices to individual get/setitems, so those should not be
boundschecked. This is not so simple, however, because parfor generates code
at the numba IR level and the boundschecking is implemented at the LLVM IR
level. So this implements it by setting a flag on the node, then globally
disabling boundschecking when that node is lowered. This is a bit hacky, but I
don't know how else to do it.

Unfortunately, this doesn't work. It fixes the test in question
(test_parfor_slice19), but then a bunch of other parfor tests fail with
'@do_scheduling_unsigned' not found in ..., and I'm not really clear why.