WIP: allow lists to have None as type
This is a very hacky first implementation of enabling a typed list with
None as the type. Since this feature hasn't been designed in from the
start it is implemented with several if statements placed at the right
position.

This is needed because the following construct, which is used in many
places in the Numba code-base has to work:

```
l = List()
[l.append(i) for i in range(10)]
```