fix typed list equals
The previous implementation incorrectly assumed that when comparing
equality of lists, we can compare the list elements for order. However,
not all possible typed can be compared for order, for example typed
dictionaries can be compared for equality but not for order and so the
previous  implementation failed during typing. This fixes that and the
new equals and not equals methods don't assume that the types can be
compared for order but for quality only.