WIP: implement disable_reflected_list switch
The 'disable_reflected_list_switch' needs to be present in both the
typing- and targetcontext. This implements it for the target context but
hacks it in for the typing context. This already makes the following
work:

```
@njit(disable_reflected_list=False)
def simple_BUILD_LIST_inference_with_append():
    l = []
    l.append(1)
    return l
```

And:

```
@njit(disable_reflected_list=True)
def simple_BUILD_LIST_inference_with_append():
    l = []
    l.append(1)
    return l
```