Improve errors / warnings for cuda vectorize decorator
The CUDA vectorize decorator provides harsh feedback to users who
provide it with unexpected arguments - an assertion failure with
no further information. This commit changes the decorator in the
following ways:

1. The nopython kwarg is now accepted, albeit with a warning that
it is redundant on CUDA. The cpu and parallel targets do also
accept this kwarg.

2. Provide a KeyError for other unsupported kwargs, consistent
with the CPU and parallel targets.

The additional tests check the expected behaviour for all targets.
For ths CPU target the decorator is tested with and without a
signature, in order to check for normal ufuncs and dynamic ufuncs.

(Minor: a small whitespace error in test_vectorize_decor.py is
also corrected)