use error handlers to cleanup api
Depending on the calling convention, different methods have to be used
to raise an exception. These different methods have different call
signatures. We unify the `error_handler` API to hold all arguments that
are shared across calling conventions and hold and differing arguments
in the state of the error handler itself. This makes the API much
cleaner across calling conventions.