Add more appropriate typing for CUDA device arrays
As reported in Issue #4628, CUDA device arrays are always typed
with 'A' layout - this is sub-optimal, as all generated code
accessing them will need to use strides passed in at launch time
instead of being hardcoded constants in the kernel. This commit
adds more appropriate typing for device arrays.

The choice of layout when typing is governed by the following
considerations:

1. The preference is to use 'C' or 'F' layout since this enables hardcoding
stride values into compiled kernels.

2. If an array is both C- and F-contiguous, prefer 'C' layout as it's the more
likely / common case.

3. If an array is broadcast then it must be typed as 'A' - using 'C' or 'F'
does not apply for broadcast arrays, because the strides, some of which will be
0, will not match those hardcoded in for 'C' or 'F' layouts.