Make device_array_like create contiguous arrays
As reported in Issue #4832, device_array_like maintains the
strides of source arrays, rather than following the behaviour of
Numpy and always creating a contiguous array. This commit changes
device_array_like so that it always creates a contiguous array.

Strides for the new array are computing using the same logic as
Numpy uses when keeping the order of the source array.