adding some documentation for the numba.future module
Adding some documentation to the numba.future module to describe what
the purpose of the module is, how to use it and how it will develop in
the future.

Also adding some documentation for the `disable_reflected_list` import
and how it affects the behaviour of Numba in regards to lists.