Don't launch ForAll kernels with 0 elements
When created with 0 elements, a `ForAll` would configure a kernel
with 0 blocks and then cause a `CudaAPIError` when attempting to
launch. This commit fixes this issue by returning early from
`ForAll.__call__` when the `ForAll` was constructed with 0
elements.

Tests are added for the original issue reported, where assigning a
zero-length slice would trigger the error, and also for directly
constructing a `ForAll` with 0 elements.