Fix various issues in CUDA documentation
- cudapysupported.rst: Remove non-existent `math.arctan`, and Python
  2-only `div` and `idiv` operators.
- device-management.rst: Fix broken references to `numba.cuda.gpus` and
  the ``DeviceList`` class.
- intrinsics.rst: Remove non-resolving reference to `numba.cuda.atomic`.
  and adjust grammar accordingly.
- decorators.py: Remove a trailing "If" in docstring and fix references
  to the `AutoJitCUDAKernel` class. Remove the type of `func_or_sig`,
  because it doesn't refer to a particular type.

This resolves most, but not all, issues in the CUDA documentation that
cause Sphinx to complain in nit-picky mode. The remainder come from
docstings in the `numba.cuda.random` module, which will require some
creativity to resolve as the types of parameters cause an issue - e.g.
they are things such as `uint64`, which is a Numba type but not a Python
class. Ideally these would link back to the table of types in the
reference manual, but some custom solution may be needed to do this in a
non-ugly way.