utils: remove backported total_ordering from functools
Uses are replaced with the standard library functools.total_ordering now
that all supported platforms (Python 3.6 onwards) provide it.

As test_utils is removed (it tested only the total_ordering backport)
the test_module test in test_runtests needed modifying since it checks
the test infrastructure by counting the number of tests in test_utils -
instead it counts the tests in test_storeslice.