fix None handling for getitem
As we now support None typed typed-lists, getitem would, in general,
have an `Optional(type)` signature, because it could return either a
value or None. But, since the type of the list is known beforehand, we
can de-optionalise getitem again. If the list is of type `item_type` it
will, if the index is valid, return an item of that type. If the list is
of type `None` it will, if the index is valid, return None.