Fix #4875: Make #2655 test with debug expect to pass
The test case TestException.test_raise_causing_warp_diverge_failing was
originally added in PR #2706 which worked around unexpected warp
divergence due to exception raising code (discussed in Issue #2655).
Since the merge of the Python 3.8 CFG refactoring the test case has
begun to pass again as it is sensitive to CFG layout.

The suggested resolution in #4875 is to remove the failing test.
However, this commit instead makes the test an expected pass as there
still seems to be some value in the test itself - this test could
provide a canary for future changes (potentially elsewhere in the
software stack) that affect scheduling or synchronisation - this test
beginning to fail again suggests something to be aware of and understand
/ investigate.