implement None-typed typed-list
This is the code for the None-typed typed-list as extracted from:

https://github.com/numba/numba/pull/4883/files

The support is very limited and was designed only for a single, very
limited use-case: populating a typed-list with a list comprehension.

Let's imagine you want to do the following:

```python

from numba import njit, typed

@njit
def foo():
    l = typed.List()
    [l.append(i) for i in range(10)]

```

While it is true, that this is fairly uncommon way to initialize a list,
it was and is used to populate a typed-list since that doesn't (at the
time of writing this) have a way to be initialised from an iterable or
similar. And so the list comprehension approach is one line shorter than
a for-loop construct.

As a by product of this idiom, is that a None-typed list is generated
and discarded as a result. If the list comprehension is to generate a
Numba typed-list, rather than a reflected list, in future, then it must
support a None-typed variant

A mentioned, the implementation here is as minimal as it can be and
doesn't have an optimizations. Most of the method are not supported and
will be guarded against. An optimized variant may not need to allocate
any memory at all, for example, since a None value will always be
returned when accessing it and we only need a count of the number of
Nones stored