enable refining the typed-list type ListType to be immutable
In order to properly test (and maybe use) the immutable typed list, we
need a way to make it immutable after creation. Otherwise it would be
impossible to actually create an immutable typed-list with content.

The implementation already has a mechanism to change the type of a list,
specifically the refinement of type. We leverage this here and simply
allow the refinement of the mutable attribute too. We add it as a
keyword argument, thereby expanding the API without breaking backwards
compatibility. So, when a list is made immutable, the type of it is
refined.

No mechanism to go back from immutable -> mutable exists yet.