Fix #5429: Exception using `.get_ipc_handle(...)`
... on array from `as_cuda_array(...)

In the implementation of `get_ipc_handle`, the owner of a MemoryPointer is
assumed to be another MemoryPointer instance (and therefore has a `handle`
attribute which points to the start of the allocation in which the
MemoryPointer's value points to). This was always true when MemoryPointers were
only created by slicing DeviceNDArrays - however, MemoryPointers can also now
have owners that are objects coming in through the CUDA array interface (e.g.
via `cuda.as_cuda_array()`).

This commit removes the assumption that the owner of a MemoryPointer is another
MemoryPointer, and instead gets the start of the allocation with
`cuMemGetAddressRange` (via the already-extant `device_extents` function).

Tests are added to test_ipc that emulate an array from another library being
used via the CUDA Array Interface. The mechanism for this test is common to
test_ipc and test_cuda_array_interface, so the `MyArray` class from
test_cuda_array_interface is moved to `numba.cuda.testing` as `ForeignArray`.

A couple of superfluous print calls are also removed from test_ipc, in the
interests of tidiness.

With this commit, the original reproducer:

```python
import cupy
import numba.cuda

a = cupy.arange(5)
a2 = numba.cuda.as_cuda_array(a)

a2.get_ipc_handle()
```

runs to completion without error.