Make _parse_{dtype,shape} public, support TypeRefs
These functions from numba.core.typing.npydecl are already used outside
the module, and will also be useful for the CUDA target - it seems they
ought not to be underscore-prefixed.

parse_dtype presently handles DtypeSpec instances, but record types may
be encountered that come in as instances of TypeRef instead - the
function is extended so that it also returns the instance type of a
TypeRef.