Fix #4809
The global tuple of uint64 is typed as a literal, but because the
members of the tuple are of type numpy.uint64, they cannot be used as
literal values (see Issue #5508). This causes a fallback to object mode.
Thanks to @stuartarchibald for identifying the issue.

Since the jump values are only used in the xoroshiro128p_jump function,
we can avoid the issue by moving the values inside the function.