Unset the memory manager after EMM Plugin tests
Setting the EMM plugin to the internal memory manager is incorrect when
the tests are being run with an external memory manager, e.g. with:

```
NUMBA_CUDA_MEMORY_MANAGER=rmm python -m numba.runtests numba.cuda.tests
```

After the EMM plugin tests execute, testing would continue with the
internal memory manager instead of the specified external memory
manager. The correct behaviour is to un-set it, so that when the new
context is created, it is set to the chosen memory manager (either the
internal one or an external one).