refine the documentation regarding index
The C implementation of the numba.typed.List assumes that any index
being passed as a parameter has been appropriately clipped an is in the
range `0 <= index < len(list)`. Any conversion of negative indices is
expected to happen at the llvm/compiler level. The function
documentation has been updated to reinforce this assumption.