fix exception during locale detection
Under certain circumstances for example if `$LANG` is unset in the
environment, `getdefaultlocale` can return the tuple `(None, None)`.
This in turn would cause an `Exception` to be raised by `join` at least
in Python 3.6, as it expects the iterable to contain `str`s. The result
was the following error message:

```
Error (locale): sequence item 0: expected str instance, NoneType found
```

Unfortunately the tests for the system info detection would then crash
making the test-suite fail. To mitigate this, we now cast everything
returned by `getdefaultlocale` to `str` to make sure that the call to
`join` will succeed. The output will then be `None.None` which is fine
as it indicates that no default locale was detected.