CUDA: Add support for CUDA 11 and Ampere (sm_80)
With CUDA 11, we need to switch to determining the supported compute
capabilities using the runtime version, because:

- CC 8.0 is supported by CUDA 11.0, and not 10.2
- The NVVM version and its IR version, as well as the soversion, are all
  constant between CUDA 10.2 and 11.0

Therefore, the existing mechanism that looks at the IR version was not
sufficient to determine if CC 8.0 is supported (and also 3.0 and 3.2
unsupported, as in CUDA 11.0). It is expected that the cudart version
will always be consistent with the NVVM version in the environment in
which Numba is used, so the runtime version is queried instead.

We need to make a couple of other small changes:

- test_get_arch_option: 'compute_30' is no longer supported by all
  supported toolkits, so we change the test slightlyso that it uses
  'compute_50' and 'compute_35', which are all supported across
  CUDA 8.0 - 11.0.
- test_warp_ops.py: We use the runtime version in _safe_skip, since the
  NVVM_VERSION variable it did use is removed.