Fix #5835: CUDA debug info is invalid
This commit moves the creation of the `CompileUnit` for `NvvmDIBuilder`
to the moment before it is added to the module. At this point, the list
of subprograms is populated, so the compile unit correctly refers to all
the subprograms, instead of referring to empty metadata, which happened
when the compile unit was created prior to any of the subprograms.

The behaviour differs between the `DIBuilder` used for the CPU target,
so a new method, `initialize()`, is added to the `AbstractDIBuilder` to
facilitate this change without breaking the `DIBuilder`.