Updates the "New Issue" behaviour to better redirect users.
This uses the config.yml functionality to add links out to discourse
in the case of users with questions/feature discussion. It also
updates the bug report and feature request templates, removing the
suggestion to ask on gitter.im as users with more question like
problems should already have been intercepted.