Change synchronization proposal to add stream
This is an attempt to codify Leo Fang's suggestion from:

    https://github.com/numba/numba/issues/4933#issuecomment-578088647

which is Option 3 in:

    https://github.com/numba/numba/pull/5162#issuecomment-635928206

This is done based on feedback in subsequent comments indicating this
may be the most promising direction to take.

A minimal implementation of the specification in Numba is added for
demonstration and discussion purposes. This implementation aims at being
correct and minimally complex, but leaves some performance on the table.