Don't make a runtime call on import
Calling `runtime.get_version()` on import initializes the CUDA driver,
even though it creates no context. This is an issue for libraries that
import Numba then set `CUDA_VISIBLE_DEVICES`, because the call at import
time already initialized the driver and fixed the set of visible
devices. This is an issue for, but not limited to, the Dask Local CUDA
Cluster.

This commit fixes this by deferring the call to `runtime.get_version()`
until the list of supported CCs is actually needed.