Fix #6260
get_ir_of_code doesn't necessarily manage to produce SSA such
that variables can be multiply defined. This leads to problems
including but not limited to nonsense IR with nodes referring to
their own label and DCE eliminating variables with the same name
that shouldn't have had the same name.

Closes #6260.