CUDA: Add support for complex power
This commit adds support for complex power using the `**`and `**`
inplace operators, and via the builtin `pow` function. The
implementation in `libdevice.py` is a translation of the logic from
`_Py_c_pow` in CPython 3.7.5.

Although `libdevice.py` seems a strange place to put this, it is more
that `libdevice.py` is strangely named - it is where the implementations
of most of the CUDA maths functions are. Note that PR #6152 moves all
these into a more appropriateley named `mathimpl.py`.

Lowering is added for `complex64` and `complex128` separately - the
`complex64` version carefully casts everything to ensure that single
precision arithmetic is used.

There is only a minor documentation update here, because the
documentation seemed to imply already that this was already supported -
however, support for the builtin function `pow` was not listed.

New tests are added that test with finite values. I did experiment with
testing with infs, nans, etc., but could find very few cases that
wouldn't raise a `RuntimeWarning` about invalid values when executing
the Python versions of the functions for comparison.

Fixes #5727.
Fixes #3386.
Fixes #2495.