Fix test_round_to_f4_halfway, andd f8 variant
The f4 variant was failing because the input value (0.5425) was not
exactly representable using a float32, and the value being used for the
test was 0.542500019, which is not halfway between 0.542 and 0.543. The
value is changed to 0.3425 to enable the test to pass.

A variant testing float64 is added, which can succeed with the value
0.5425 as input when rounding to three digits.