Adjust tolerance for math tests with float64 dtype
The lgamma implementation in Python differs to that in GNU libc to the
point that the results can differ relatively by as much as ~4.22e-14.
This causes a fail with cudasim, which is rectified by reducing the rtol
down to 1e-13. This is still far tighter than the rtol originally used
in these tests, which was 1e-5.