Update bug and feature request templates.
Wording changes to try and push general questions onto discourse and
keep github issues/PRs for development. This makes bug and feature
request text point to "below" which has a General Question category
as present on discourse. Also changes the bug report template to ask
for a self contained reproducer with a simple definition of what that
is.