Detect when an array variable is multiply defined with inconsistent shapes/sizes. Add such arrays to the multiply defined set and if we found any then run array analysis again so that equivalences insert during the first pass before we saw that the array was multiply defined can be eliminated.
Add some checks in some places to make sure that certain functions are only executed for arrays and not scalars.

For multiply defined arrays, still create size vars but use the same size var
for every assignment to that array and in combination with the above those size vars
should not become an equivalency for anything before the assignment but can be afterwards,
again at the block level.