Fix #6556 extend variable lifetime to block end.
Potential fix for #6556, this delays issuing `ir.Del` nodes for
dead variables until the end of the block in which they become
dead. The result is that variable lifetime is slightly extended but
it prevents use after free issues from too early decref when e.g.
`.ctypes` or FFI buffers are used with arrays.

Fixes #6556