Prevent load of libnvvm.so at import time
The previous commit caused libnvvm.so to be loaded at import time for
various reasons, some of which could be resolved by reworking imports in
the CUDA target, but the @cuda.jit-decoration of functions in
numba.cuda.intrinsic_wrapper would force the load of libnvvm.so through
an eventual call to compile_device_template. compile_device_template
imports the CUDA target descriptor, which creates an empty module and
needs to load NVVM to determine which version it is for the metadata
addition.

This commit resolves the issue by modifying the CUDA target descriptor
such that it only initializes the typing and target contexts when they
are required - compile_device_template only needs the typing context, so
the target context initialization (and therefore the load of libnvvm.so)
are avoided at import time. The modifications to the descriptor also
bring it into line with the idiom used for the CPU and ufunc targets,
which also construct a single instance of the target class rather than
having class variables for the typing and target contexts.

Unfortunately it is not yet possible to move the imports of the CUDA
target descriptor up to the module level in numba.cuda.compiler, as this
creates a circular import. This may be solvable with more effort :-)