Don't parse the IR for CUDA modules with llvmlite
In preparation for using the bitcode interface to NVVM, this commit
avoids parsing the IR for CUDA modules with llvmlite then dumping them
back to a string - the rationale for this is that the compilation flow
would have involved sending the IR through three different versions of
LLVM:

1. LLVM 10 via llvmlite for parsing and dumping in the CodeLibrary
2. LLVM 7.0.0 to convert the LLVM 7 IR to LLVM 7 bitcode
3. LLVM 7.0.1 in NVVM

Removing the first step aims to reduce complexity and the potential for
subtle issues, especially as llvmlite moves ahead to newer LLVM
versions.

To support this change, the `CodeLibrary` is refactored slightly to
split out a `CPUCodeLibrary` class whose methods still implement the
first stage above, which is required for the CPU target.

A corresponding refactoring of the `BaseCPUCodegen` class is also made -
the base class shared between CUDA and the CPU is `Codegen`. This cleans
up the abstractions a little by eliding the need for the CUDA codegen to
override some of the CPU codegen's methods to "hide" them.

This change mostly, but does not completely avoid parsing IR with LLVM
10 through llvmlite when using CUDA - some parsing is still required for
`BaseContext.get_abi_sizeof`, which parses types to find out their size.
This is expected to be both mostly harmless, and difficult to avoid, so
it is left in place.

Some other remarks on the changes:

- The 32-bit CUDA triple is removed, since it is no longer supported.
- This commit also deprecates the `nvvm_options` kwarg for `inspect_ptx`
  as it seems to support inspecting PTX generated with options different
  to those that would normally be used when code is JIT-compiled, and it
  will hamper further planned reorganisation of the way code generation
  works in the CUDA target. It is not expected that it is widely used.
- The syntax for various stub functions (`___numba_atomic_double_add`
  etc.) is a little different coming from llvmlite compared to being
  dumped from LLVM 10, so the replacement strings are updated to account
  for this.
- llvmlite generates debuginfo flags in the old sttle (e.g.
  `isDefinition: true` instead of `DISPFlagDefinition`) so this
  substitution is no longer required.