Remove explicit external linkage for kernels
The default linkage is external already.

When setting the linkage of functions in linked modules, the check for
external linkage is removed - as illustrated in the comment above this
code, all linked modules are device functions that would need
linkonce_odr, so there is no need to check their linkage.