CUDA: Remove envvars.py and NumbaPro envvar warnings
The `NUMBAPRO_*` env vars have been deprecated and ignored since 0.45 -
this commit removes the warning about them. Now the warning is gone, the
`get_numba_envvar` function serves no purpose above normal environment
variables, so the variables that this was used for are converted to
normal config variables, and documented.

I changed the `VERBOSE_CU_JIT_LOG` variable name to
`CUDA_VERBOSE_JIT_LOG` for consistency - I don't believe this needs a
deprecation cycle as it's never been public (and is likely unused in
general).