CUDA: Fix parallel testing for all testsuite submodules
Running the testsuite in parallel for the `numba.cuda.tests` module
works already, but running it in parallel for submodules, such as
`numba.cuda.tests.cudapy` did not. This is because the
`numba.cuda.tests.__init__.load_tests()` function ensured that the list
of supported compute capabilities in `numba.cuda.cudadrv.nvvm` was
initialized in the parent process, but the `load_tests()` function in
its child modules did not.  This commit rectifies this issue by ensuring
that it is initialized for each of the submodules that might want to use
NVVM in a test that can be run in parallel (i.e. not a `CUDATestCase`
requiring a context).

Some other issues were discovered and are fixed:

- In nvvm.py, if the CUDA runtime was not present,
  `cudart_version_minor` was not set, so formatting `ctk_ver` raised an
  exception, leading to confusing error messages. It is now set to 0 for
  an absent toolkit.
- `TestNvvmWithoutCuda` had the `SerialMixin`, which shouldn't be needed
  for a test that is not supposed to require a context. The mixin is now
  removed.
- `TestNvvmWithoutCuda.test_nvvm_llvm_to_ptx` actually does require a
   context, because it compiles a kernel (which will be linked and
   loaded, etc.)! This test just shouldn't be in a test class that is
   not supposed to require a context, and is probably left over from a
   time when binding (linking and loading) did not occur automatically
   (many years ago). Since there are better tests of `nvvm_to_ptx` in
   the `cudapy` tests, this test is removed.

Running the whole CUDA test suite or its submodules in parallel with the
`-m` flag, e.g.:

    `python -m numba.runtests numba.cuda.tests.<module>`

now works correctly for each of the modules:

- No submodule, just `numba.cuda.tests`
- `cudadrv`
- `cudapy`
- `nocuda`
- `cudasim`
- `doc_examples`

(verified manually)

Fixes #6703.