CUDA: Support for `@overload`
Joint work with @stuartarchibald.

The implementation consists of:

- Addition of `get_call_template`, `nopython_signatures`, and
  `get_overload` to the class that is now `DeviceDispatcher`. These are
  needed by the overload infrastructure, and are inspired by the
  equivalents in `_DispatcherBase`.
- Registration of a CUDA `jit` decorator. The registration with
  `dispatcher_register` is removed and replaced with this new
  registration, because it did nothing (it was for the `target` kwarg to
  the `numba.jit` decorator, which is deprecated). The `jit` decorator
  for overloads is a bit special in that it always adds the
  `device=True` kwarg to the regular `jit` decorator, because overloads
  will always be device functions.
- Addition of a `no_cpython_wrapper` kwarg to `jitdevice`, which is
  needed because the overload infrastructure passes it. No action is
  required regardless of its value, because there is no CPython wrapper
  in the CUDA target.

A refactoring is also made: `DeviceFunctionTemplate` is renamed to
`DeviceDispatcher`, as it reflects more closely what it is - it is
analagous to the `Dispatcher` class used for kernels, and the
`DeviceFunction` class is more analagous to the `_Kernel` class.
Eventually `Dispatcher` will subsume these classes as we move towards a
more "modern" dispatcher implementation in the CUDA target.

Tests are added; these are all skipped on the simulator because
overloading doesn't really exist as a concept without compilation.