Use value-based check for CUDA overload tests
Using prints and capturing stdout to test overloads is both overly
complicated and difficult to debug (terminal-based debuggers and
debug prints don't work during the tests). The value-based check that
replaces it suffers neither of these issues, and it is at least as clear
what each test does (more so, in my opinion).