Use correct CUDA driver library on WSL
Using the wrong libcuda.so on WSL can result in `CUDA_ERROR_NO_DEVICE`
being returned on the call to `cuInit()`. The correct library location
seems to be `/usr/lib/wsl/lib`. This PR adds this folder ahead of the
usual library paths for the CUDA driver, so that on WSL systems it is
detected ahead of the other libraries.

Fixes #7104.