Reinstate and deprecate eager compilation of device functions
Removing eager compilation of device functions without a deprecation
cycle could break external user code. This commit reinstates eager
compilation of device functions, and deprecates it with a recommendation
to use `compile_ptx()` instead (the "legitimate" use case for eager
compilation of a device function is to compile a Python function to PTX
for use outside of Numba, but this is better served by `compile_ptx()`.

Test results are now:

```
Ran 1178 tests in 88.726s

FAILED (failures=2, errors=1, skipped=10, expected failures=1)
```