CUDA: Skip IPC tests on ARM
These are unsupported on Tegra platforms (such as Jetson). There are
some CUDA ARM systems that support IPC, but for now we will count these
as not supported by Numba, and skip the tests regardless.

In the longer term (when we get to test Numba on CUDA ARM systems that
support IPC) this should be re-visited, either by finding a more precise
check for skipping the tests than just "Is the CPU architecture ARM?",
or by moving the IPC implementation to use `cuMemCreate()` and obtaining
shareable memory handles, which will be supported on Tegra.

Fixes #7178.