CUDA: Warn when debug=True and opt=True
As per Issue #7214, using debug=True with opt=True can cause NVVM to
crash (or the link can fail with an error message from ptxas). We can't
easily stop the user from doing this by default as it might break a lot
of existing code (particularly that which uses debug=True to force a
check for exceptions) but we can warn them.

This commit adds a warning when debug and opt are both true, and some
tests for most cases. The debuginfo tests are also updated so that they
don't trigger the warning and make the test suite noisy. However, I
chose not to "fix" the tests that check for a raised exception (e.g. in
test_exception and test_fastmath) - I think the test suite should be
emitting warnings there to highlight/remind about the fact that we need
to split debug=True into at least two options (e.g. debuginfo=True and
exceptions=True).

Fixes #7214.