CUDA: Add libcuda.so to libs.test()
It is useful to show the driver location and test loading it in
libs.test() - this commit adds this functionality, with a couple of
limitations:

- The list of candidates can be shown, but it is not easy to state
  the exact path of all candidates, because the loader's search order is
  not easy to access.
- If the loading succeeds, the exact path of the loaded library may not
  be visible (e.g. it may just show that loading succeeded from
  "libcuda.so").

That said, this should give some insight into the loading of the library
and whether it succeeds, so this seems worthwhile for now - perhaps more
sophisticated interaction with the OS to determine exact locations could
be added later.

Some modification to the `find_driver()` function in driver.py was
needed to expose the relevant information to `libs.test()`.