Fix #6001: typed.List ignores ctor arguments with JIT disabled
When `List.__new__()` is called, the constructor arguments get swallowed
by `lsttype`, which is the first argument (after the class) and is also
a keyword argument. This doesn't matter when JIT is enabled, because
`__new__()` returns an instance of `typed.List`, so the interpreter then
calls `__init__()` on the result.

However, when JIT is disabled, `List.__new__()` returns an instance of
`list`, and therefore the interpreter does not call `__init__()` on it
at all - it is the responsibility of `__new__()` to initialize the
returned object.

(See https://docs.python.org/3/reference/datamodel.html#object.__new__
for an explanation of the behaviour and requirements described).

To implement correct behaviour with JIT disabled, we need two changes:

- The non-keyword arguments need to be placed ahead of keyword arguments
  in the parameters to `List.__new__()` (`*args`).
- When JIT is disabled, in addition to creating a new `list`, we also
  explicitly call its `__init__()` method with our `args` and `kwargs`.

Fixes #6001.