Use normal list constructor in List.__new__()
Calling `list.__new__()` without arguments and then `list.__init__()`
with arguments on the resulting object is not guaranteed to work - the
fact that `list.__new__()` doesn't refer to its arguments (and therefore
doesn't strictly require them to be the same as the arguments to
`__init__()`) is an implementation detail.

It is safer to just call `list(*args, **kwargs)` instead.