Support typed lists in `heapq`
The `heapq` implementation checks for reflected lists, which bans its
use with typed lists. This commit modifies the check so that it also
allows typed lists. There are no further functional changes required for
correct operation with typed lists - the `heapq` tests are updated to
run all tests with reflected lists and with typed lists. The
modifications to individual tests all convert to/from the chosen list
class (either CPython lists for reflected lists, or typed Lists). In
some cases typed lists need to be converted back to Python lists for
functions like `assertPreciseEqual` to compare them with Python lists
used in the reference (CPython) implementations of the functions.

I have checked that the tests with typed lists never use a reflected
list by running the test class individually - no warnings about the use
of deprecated reflected lists are emitted when running it:

```
$ python -m numba.runtests numba.tests.test_heapq.TestHeapqTypedList -m
Parallel: 21. Serial: 0
.....................
----------------------------------------------------------------------
Ran 21 tests in 16.737s

OK
```