Add size_t and ssize_t types
These are mainly for use with `cfunc`, to match the defined signature of
a C API. In flat-memory ABIs these are the same as uintptr_t and
intptr_t, but's not guaranteed to hold forever. More practically, having
these names will make code clearer, as it will match the C names for the
interface being implemented, and make it clear that a parameter is
holding a size/count rather than a bit-cast pointer.