CUDA: Fix potential leaks when initialization fails
When CUDA driver initialization fails, the driver singleton object
persists holding on to an exception object that references calling
frames and potentially other objects. This can create a leak in the case
where modules that attempt to initialize the driver and then fail are
created and destroyed.

This change rectifies the issue by holding on to a string describing the
error instead of the exception object. There are a couple of small
related changes:

- `initialize()` is changed to `ensure_initialized()`, and the caller no
  longer needs to check whether it should be called - it can always call
  it when it needs to ensure that the driver is initialized.
- `cuda.cuda_error()` returns the error string instead of an exception
  object. Constructing an exception object here just to maintain the
  original behavior seems a bit convoluted; it's likely that any code
  using `cuda_error()` is checking whether its return value is `None`
  rather than looking for a specific instance of an exception class to
  see if an exception occurred.

Some initialization tests are added - for the failing cases we need to
run in a subprocess to avoid interfering with the initialization of the
driver in the process in which we're actually running tests. The failure
of `cuInit(0)` is accomplished by a slightly unorthodox patching of
`driver.cuInit`, which is needed because driver functions are added to
the `Driver` object on-demand, so there is nothing for
`mock.patch.object()` to replace at the time we need to set up the mock.