cleanup Numba metadata
Cleanup the author and email listing on setup.py, the mailinglist is no
longer used and the company has been renamed. Also, we now host our docs
on readthedocs, so update that too.

In addition remove the mention of the "old" way of contact from the
README and the Sphinx based documentation.