CUDA: Provide stream in async_done result
It's awkward to await on a `Stream.async_done()` for the purpose of
enqueuing something new onto the stream, because there is no way to tell
which stream finished from the completed future - this forces the user
to maintain a mapping between futures and streams.

This commit makes the result of the future be the stream on which
operations completed, so that the user can obtain the stream from the
completed future.