Fix sentry_record_alignment for NestedArrays
Now that the backend data type of nested arrays is a normal array
structure, the check for alignment of nested arrays within a record
needs to look at the type of elements of the nested array when
considering alignment, instead of considering the array structure type's
alignment. This fixes the CUDA test failures introduced by earlier
commits in PR #7359.