Update the docs and error message for errors when importing Numba.
This updates the error message users see when the import of Numba
fails. It directs users to new documentation that outlines common
issues and has suggested fixes. This error is now considered
non-reportable as pretty much all occurences are due to one of the
few cases that are now documented. Should users have an issue that
is not covered, the docs direct them to the discourse forum to ask
for assistance.