Move the type annotation pass to post legalization.
This PR moves the type annotation pass to run after legalization,
this is so as to pick up the variable lifetimes correctly. The
`AnnotateTypes` pass now depends on the `IRLegalization` pass. The
rest of the changes are fixes to change use of members of the type
annotations object (e.g. `typemap`) to the same member on the
pipeline `state` object.