Use omp_set_max_active_levels() instead of omp_set_nested()
omp_set_nested() is deprecated and often results in a warning on recent
installations. Following the discussion in Issue #5275, this commit
changes the call to omp_set_max_active_levels(). Ideally we'd set it to
the maximum supported nesting level, but we can't determine what that
is; the function `omp_get_supported_active_levels()` was only added in
OpenMP 5.0, which is quite recent and was not supported by some
implementations until very recently (e.g. it was added to GNU OpenMP in
mid-2020). We arbitrarily set the nesting level to 32, which is far
greater than any real requirement, and the implementation will limit the
nesting levels to the supported number as per the spec if it is less
than 32.