Simplify nested array assignment lowering + add tests
This simplifies the lowering of a nested array assignment by explicitly
copying the data from the RHS to the nested array's data.

A test for an assignment of a 2D nested array is also added - since the
new copying code relies on using `nitems` from the element type, this
seemed prudent to add (though it is expected that `nitems` does of
course always equal the total number of elements of any nested array).