dissallow llvm-openmp 12.* for M1
The `llvm-openmp` package has a Numba relevant bug that only manifests
when using that package on M1 silicon (new Apple chips) in version 12.
This will allow building Numba (the headers are needed) but dissallow
running it with `llvm-openmp` in version 12.* on Apple M1 silicon.

The `llvm-openmp` bug is here:

https://github.com/llvm/llvm-project/issues/49923

This will also be future proof and allow `llvm-openmp` to be installed
in version 13 as soon as that becomes available from `main/defaults of
if people want to use `conda-forge`, since that already has version 13.