remove the roadmap from the sphinx based docs
As discovered in https://github.com/numba/numba/issues/7930 the roadmap
isn't being actively maintained. This may cause confusion and create the
impression that the project is no longer active. So for now, no
information is better than outdated

If any of these items are still relevant and need to be done they ought
to be converted to tickets.