disassemble_cubin: Don't check filename of FileNotFoundError
The filename may not be present on Windows, and it's not clear that
checking the filename adds anything, given that there is only a single
line of code in the try block with a fixed filename.

When the filename is empty, no exception is raised by
disassemble_cubin(), and it then attempts to access cp, which has not
been set. To avoid this issue, we now unconditionally raise the
RuntimeError that indicates that nvdisasm is not present.