Support "static" __getitem__ on Numba types in `@njit` code.
This adds support for the `__getitem__` operation on Numba types
within `@njit` compiled regions. i.e. it makes this sort of thing
possible:

```python
from numba import njit, types, typed

@njit
def foo():
    return typed.List.empty_list(types.float64[:, ::1])

foo()
```

The logic in the Type metaclass is relatively relaxed over what's
acceptable in a `__getitem__`, this patch is necessarily more
restrictive and the compilable `__getitem__` is implemented only
for types that are practical for this sort of use i.e. those in
the numerical type system.

Tests are added and docs are updated to note that this is now
supported.

Closes #8808
Closes #7683
Closes #5748