Remove coverage script
Due to various slight oddities in `numba._runtests._main()`'s handling
of arguments, it's actually hard to make this script work robustly
across platforms and different invocation methods. Instead it would be
better to document how to use the `coverage` CLI tool, as is done in
practice by maintainers when investigating coverage.