CUDA: Remove context query in launch config
This swaps the context query in the launch configuration for a
statically-chosen threshold of 128 blocks, removing a source of overhead
in kernel configuration.

This also fixes a failure in
TestWarnings.test_efficient_launch_configuration, which fails with:

```
======================================================================
FAIL: test_efficient_launch_configuration (numba.cuda.tests.cudapy.test_warning.TestWarnings)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/swqa/NUMBA/envs/NumbaGPUTest/lib/python3.9/site-packages/numba/cuda/tests/cudapy/test_warning.py", line 33, in test_efficient_launch_configuration
    self.assertEqual(len(w), 0)
AssertionError: 1 != 0
```

when run on a device with 132 SMs - this is because the test launched
256 blocks, but the check for emitting the warning requires 2 * (SM
count), which is 268 (the test was written assuming such a large GPU did
not exist).