Make the NRT use the "unsafe" allocation API by default.
This patch adds to the NRT C API to create equivalent "safe" and
"unsafe" functions for allocation associated with meminfo. The
NRT context is updated to only use the "safe" functions if
`NRT_DEBUG` is enabled. This is to increase performance by only
calling `memset(3)` to generate debug markers at allocation time
when in `NRT_DEBUG` mode. The "aligned" allocators are updated to
calculate alignment offsets avoiding modulo if the alignment is
specified as a power of 2.

Fixes #7887