CUDA: Remove note about toolkit version from CG docs
This comment was only relevant when there was the possibility of
cudadevrt being missing for an installation, when (broken) older CUDA
toolkit packages that happened to be missing it were still supported.
Now they are unsupported, cudadevrt can be assumed to be present so
there is no need to explicitly mention it in the CG docs.