Revert #8265.
This reverts #8265 as it causes mainline builds to fail in parfors
testing. The cause seems to be related to scope transfer for the
gufunc and resulting name collisions in the type map during array
analysis as a result. Tracking issue #8288.

This reverts commit 6f702299ba2cbe66a196b05a43a49d06bc529008, reversing
changes made to e3d4912f8e6b1ddf8f7b42d73b963c18595d8824.