CUDA: Fix test_const_array_3d for NVVM34
Now that we use the standard address space casts, even NVVM32 manages to
load a complex number as a vector of 2x f32 (the old conversion
intrinsic may have been a barrier to some analysis / transformation
required to do this). This commit updates the test to always expect a
vector of two floats to be used to load the complex.