CUDA: Support for multiple signatures
When a list of signatures is passed to the `@cuda.jit` decorator, the
dispatcher is compiled for each signature. This does not result in a
specialized kernel, even if the list contained just one signature,
though compilation will be disabled for the dispatcher.

This ensures that:

- An exact match is required when a list of signatures is provided
  (matching the dispatch behaviour of the CPU target)
- An exact match is required, but not checked for, when a single
  signature is provided (matching the current CUDA target behaviour).
- Other functionality that depends on whether a kernel is specalialized
  (getting shared memory per block, for example) retains its existing
  behaviour for specialized kernels, such as returning the shared memory
  per block directly instead of in a dict for each implementation. For
  dispatchers where multiple signatures were provided, a dict should be
  returned by these functions.

Tests for multiple signatures are ported from the CPU multiple signature
tests in `numba.tests.test_dispatcher`. Note that there is a difference
betweem expectations for kernels and device functions when ambiguity
cannot be resolved for an overload - launching a kernel with ambiguous
overloads is not permitted, but compiling a call to a device function
with ambiguous overloads is permitted - this feels a little odd but is
the way Numba currently seems to work for dispatch vs. function calls,
and is recorded separately as Issue #8307.