Revert "Move array '.imag' and '.real' to overloads"
This partially reverts 0401550bc7a9e98add3e7d7728c193943d77d75c.

The changes to array `.real` and `.imag` are undone - the implementation
of `.imag` for real arrays looks a little different than the original
version because the helper function it used, `numpy_zeros_like_nd()`,
no longer exists. However, the code in this branch implements the
equivalent operation of creating an empty array and filling it with
zeroes.

The fixups that were part of the reverted commit are not undone, because
they are necessary for correctness and don't have any negative
side-effects.