CUDA simulator: fix `array_core` for NumPy 1.23
Since numpy/numpy#21477 the strides of a 0-sized array are 0 instead of
the itemsize. This causes `array_core()` to erroneously attempt to index
the array with numeric indices instead of empty slices.

Since this is a special case (empty arrays), a simple solution seems to
be for `array_core()` to return the array itself when the array is
0-sized. In order to maintain consistency of the `array_core()` function
between the simulator and hardware targets, we make this change to both
implementations.