Shorten the time to verify test discovery.
The Numba project uses `python -m numba.tests.test_runtests`
in build scripts to verify that test discovery and the test suite
itself is working as expected. This is a guard to make sure that
when the test suite is run, it's actually running what is expected.

This patch tries to shorten the time it takes to get to the point
of the test suite being validated. It does two things:

1. Hoists a costly loop invariant from a loop in one of the tests
   in the test_runtests module.
2. Caches the mtime of test files based on class as this should be
   invariant (gains about 3-4% improvement on runtests.py -l).