remove GitHub username from changelog
This modifies the changelog such that the configured username isn't
used. Only the user login is used.

If the login is `maxm` and the configured username is `Max Mustermann`,
the changelog entry would have contained:

```
(`Max Mustermann  <https://github.com/maxm>`_)
```

And with this change the entry would now contain:

```
(`maxm  <https://github.com/maxm>`_)
```