Adjust stencil tests to use hardcoded python source opposed to AST.
This patch comprises largely "mechanical" changes, the aim is to
drop reliance on the Python 3.7 AST ahead of being able to drop
Python 3.7 support entirely. To do this, all the use of the AST
and manipulations of it have been replaced with hardcoded
equivalent pure python functions. This diff touches almost the
entire test file and so fixes to the remaining few flake8
violations were rolled into the patch at the same time.