Changes how tests are split between test instances
Since tests are dynamically generated on different platforms, uniformly slicing
a list of tests can cause individual tests to be lost or tested in multiple
platforms. This changes tests to be sharded rather than sliced so that each
worker is independently making the same choice on a per-test basis and all
dynamically generated tests are always run.