CUDA: Reduce memory pressure from local memory tests
Local memory usage requires a reservation of (bytes per thread) *
(number of SMs) * (Max resident threads per SM), which for kernels with
local memory requirements of 80KB per thread (as is the case for these
tests) forces allocations into the low GBs - this is a problem on the
buildfarm as multiple tests run concurrently and can result in
out-of-memory conditions.

This commit reduces the local memory allocations by a factor of 10 - I
worry that reducing by another order of magnitude might cause local
memory not to be allocated in some cases.

Some copy/paste errors that slipped through in these tests are also
resolved (including the name of one of the tests!).