CUDA: Fix supported CC logic
The logic for determining the supported compute capabilities was
completely wrong, and needed updating in a complicated way for each new
toolkit. This replaces the incorrect logic with a simpler mechanism:

- We have a list of compute capabilities that we're aware of.
- We have the minimum and maximum supported CC for each toolkit.
- We generate the list of supported CCs as the range of known CCs within
  the min/max for the current toolkit.

For unknown toolkits, all we can do is assume that it supports all the
toolkits that we know about that aren't deprecated upstream (i.e. those
from 5.3 onwards).

This should be much easier to update - for each new toolkit, we just add
a new entry in the mapping from toolkit versions to min/max supported CC
versions. If we drop support for a toolkit, we can just delete it from
the mapping.