CUDA tests: always add the NVVM IR version
The NVVM IR version for IR given to NVVM should always match the version
that it supports - some versions of NVVM can reject IR where the version
doesn't match.

When the version isn't specified, it is implicitly 1.0, which is why
this commit mostly consists of additions that add the IR version to
modules.