CUDA: Fix inspect_llvm()
Only the wrapper kernel IR was being returned by inspect_llvm() - it
should return the IR for all modules, joining them together in a
similar manner to how PTX code is joined for inspect_asm(). The
underlying implementation is in the CUDACodeLibrary, so it is the
get_llvm_str() we need to fix.