Fix CUDA lineinfo issues
Line info emission suffered from two problems:

1. It was missing for device functions.
2. Asking Numba to generate lineinfo for kernels resulted in full debug
   info and deoptimizations (many redundant moves in the generated SASS,
   for example) being turned on.

The first issue was probably introduced by a device function /
dispatcher refactor, that resulted in the lineinfo flag getting lost,
and has a simple fix: we ensure that `lineinfo` is passed to
`compile_cuda()` as appropriate in `Dispatcher.compile_device()`. This
regression (or perhaps an issue from the beginning of lineinfo support)
was able to exist because there was no test of lineinfo in device
functions, so extra testing is added to ensure it remains correct.

The second issue was caused by the `emissionKind` of the debug info. As
well as the decision to emit debug info being made based on its presence
or absence in the LLVM IR, the choice between full debug info (and
deoptimization) or just lineinfo directives is also controlled by what
appears in the IR. When the `emissionKind` is `FullDebug`, then full
debug info is emitted and deoptimized code is generated. To generate
lineinfo only, the `emissionKind` needs to be `DebugDirectivesOnly` for
PTX generation (this is common between both NVVM and the upstream NVPTX
target). The change in LLVM that introduced this is
https://reviews.llvm.org/D46021.

To implement control of the `emissionKind`, the following changes are
made:

- Changing the `emissionKind` to be a `DIToken`. This is not strictly
  necessary right now, as we could still emit integer values (e.g. 1 for
  `FullDebug` or 3 for `DebugDirectivesOnly`, but it makes it harder to
  infer what is going on to a casual reader of the emitted IR.
- Addition of a new flag `dbg_directives_only` to control whether only
  the `.loc` and `.file` directives are emitted, as opposed to full
  debug info. This flag is only enabled by the CUDA target when
  `lineinfo` is `True`.
- Only disabling SROA-like passes when debuginfo is enabled but not when
  debug directives only are to be emitted - this aims towards keeping
  the code generation unmodified when generating line info only.
- Various changes needed to propagate the `dbg_directives_only`
  information into the right places as appropriate.

Testing for lineinfo in the CUDA target was a bit weak so it is
augmented with this commit to ensure it stays working. The additions
check both the LLVM IR metadata and PTX, and also check the debug info
for both kernels and device functions.

Fixes #8581 - "CUDA: Lineinfo doesn't work for device functions"
Fixes #8582 - "CUDA: Enabling lineinfo causes many redundant moves in
               SASS"