fix failure of test_cache_invalidate due to read-only install
In certain circumstances, such as installation through the Nix package
manager, Numba's source code is marked read-only. When the cache test
file is copied using `shutil.copy`, the destination ends up read-only as
that function copies file permissions too. This causes
`test_cache_invalidate` to fail because it cannot write to the file to
attempt to invalidate it. Fix the issue by ensuring the file is
read-write after the copy.