Prevent use of NumPy's MaskedArray.
`MaskedArray`s are not supported by Numba. They "work" in the
case of just passing them through a JIT compiled function as
they are sufficiently similar to an `np.ndarray` for that to
work. However, operations using MaskedArray's are likely to be
performed incorrectly as there is no underlying support in
Numba for them.

This patch:

* Catches the use of a MaskedArray and raises an exception.
* Adds a test for the above.
* Adds a note to the docs about the lack of support.