CUDA: Run Dispatcher tests on simulator where possible (+ fixes)
All Dispatcher tests were skipped on the simulator, but some parts of
the tests can and should run (particularly those checking multiple
signatures). This commit refactors the test to run all those that apply
to the simulator on it.

This picked up a couple of errors in the simulator that are also fixed:

- Providing a list of signatures confused the simulator's `@jit`
  decorator, because it didn't anticipate a list of signatures.
- The docstrings for jitted functions didn't match the underlying
  function in the simulator.