Handle unknown ufunc methods gracefully
`init_ufunc_dispatch()` should set an exception if it's going to fail.
Without setting an exception, an unrecognized ufunc method (e.g.
`resolve_dtypes()` in NumPy 1.24) results in:

```
SystemError: initialization of _internal failed without raising an
exception
```

This commit modifies it to set an exception notifying the user of the
error. Now when an unrecognized ufunc method is encountered, the
traceback ends in:

```
    from numba.np.ufunc import _internal
RuntimeError: Unexpected ufunc method resolve_dtypes()
```