test_ufuncs.py: Replace CompilationCache and compile_isolated
All compilations now use the jit / njit decorators.

This preserves test semantics - a minor change had to be made to ensure
that datetime and timedelta ufunc tests had matching types between the
arguments provided and those for which the functions were compiled -
previously types with and without units were being mixed, which was OK
with compile_isolated because there was no type checking, but is not OK
when the jit decorator is checking types in eager compilation mode.