Add pop NULL to CALL_FUNCTION_EX byteflow analysis.
The cPython compiler seems to add a PUSH_NULL ahead of a CALL-like
instruction irrespective of whether it will be used. This causes
confusion in the byteflow analysis in the case of a loop containing
a CALL-like that doesn't consume the NULL as the state at the end
of the loop has the NULL on the stack and so it appears to be a
"new" state with the same PC offset.