NumPy typing: expose lists of ufuncs and registration function
This prepares for implementing ufuncs in CUDA, by exposing the lists of
supported ufuncs in `npydecl.py` and the function for registering them,
now called `register_numpy_ufunc()`. This function is modified so that
it accepts a method to register the ufuncs with, so that it can either
register them with the CPU target (the default) or another target (whose
register function is passed in).