NumPy lowering: Allow ufunc registration with custom registry
This prepares for implementing ufuncs in CUDA by adding a parameter to
`register_ufuncs()` that accepts the method to register the lowering, so
that it can either register them with the CPU target (the default) or
another target, whose lowering registation is passed in.

This is a similar purpose to the previous commit, but for lowering
rather than typing.