UFunc tests: Additional class in TestUFuncs hierarchy
In order to reuse the UFunc testing machinery for CUDA, we introduce an
additional class into the `TestUFuncs` class hierarchy,
`BasicUFuncTest`. This class defines a `_make_ufunc_usecase` method that
can be overridden by the CUDA target, so that is can generate a CUDA
usecase rather than the default CPU one.