CUDA: Add function to get implementation for function on given types
In `mathimpl.py`, the lowering of `math` module functions is defined. We
need to reuse these implementations for CUDA UFuncs (until there is a
Numba-wide `libm` implementation available) so we add two functions that
provide a way to get the implementations for a given function and types.
The functionality of `mathimpl` is otherwise unchanged.