Make Numba dependency check run ahead of Numba internal imports.
This patch makes the runtime check that ensures the Python,
NumPy and SciPy verions are accepted by Numba occur before
importing any of Numba's internal modules. This is to prevent a
case where one of Numba's modules is incompatible with some version
of a critical dependency but the error mode is e.g. SystemError,
this preventing the correct reporting of the actual problem to the
user.

Fixes #8718