CUDA: Fix performance regression when lineinfo is enabled
Enabling lineinfo on CUDA would result in correct lineinfo being
generated but introduced performance regressions by preventing inlining
of device functions (reported by @fs-nv).

This commit rectifies the issue by allowing inlining to be performed
when only lineinfo is requested. The actual change is made to the code
that was in `DIBuilder.mark_subprogram()` but this is moved to
`BaseLower.pre_lower()` - the original position of this code violated
reasonable abstraction because it meant that calling
`DIBuilder.mark_subprogram()` actually changed the code being marked.

This requires corresponding changes to
`test_lineinfo_in_device_function()`. The test was looking for lineinfo
in a separate device function, but instead it needs to search for
lineinfo pertaining to an inlined device function. Additionally, it also
disallowed the emission of any section named `.debug` - however, when
device functions are inlined, a section named `.debug_str` is emitted to
hold the names of the inlined functions. Therefore, we instead disallow
a section named `.debug_info`.