Stale bot: Use `abandoned - stale` label for closed PRs
Using the `abandoned` label for stale-closed PRs causes the bot to
remove `abandoned` from PRs that have it, because it plans to add it
back when it closes them.

To prevent it vandalizing other PRs labelled `abandoned`, and to
distinguish them from auto-closed PRs, this change sets it to use
`abandoned - stale` to label auto-closed PRs instead. The `abandoned -
stale` label has already been created:
https://github.com/numba/numba/pulls?q=is%3Aopen+is%3Apr+label%3A%22abandoned+-+stale%22