CUDA: Don't explicitly set targetoptions in the jit decorator
The aim of this change is to make it possible to determine which options
were set explictly by the user, and which were left unset, so that
inheritance of options can be implemented later.

The problem with setting them in the decorator is that that it inserts
the names of options into the keys of targetoptions, which made it
impossible to tell just by looking at the keys if the user set them.