Move the minimum supported NumPy version to 1.21
This moves the minimum supported NumPy version to 1.21. The patch
removes code specialising on any version of NumPy prior to 1.21
(the oldest of which was from 1.16). It also updates code to remove
workarounds necessary due to changes in the NumPy API (e.g.
np.broadcast_shapes). Docs, buildscript and `setup.py`/`__init__.py`
are also updated.

Closes #8707