CUDA math implementations: special case tanh for lookup
When looking up the implementation to underly a ufunc, the `unarys` list
is searched. However, `tanh` doesn't appear in this list because it has
a different implementation that specialises based on the compute
capability of the target. In order to locate the implementation when
looking it up, we add it to the list we search in
`get_unary_impl_for_fn_and_ty()`.