Fix broken np.allclose, add np.isclose
It was previously impossible to use `np.allclose` as the default
arguments `rtol`, `atol`, and ´equal_nan` had types `float` and `bool`,
whereas the code was testing for the Numba types `Float` and `Boolean`.

Also replaced `np.isclose`, the element-wise equivalent of `np.allclose`.

It is also debatable whether `allclose` should be an array method; it is
not in numpy.