Don't return the function in CallConv.decorate_function()
This function mutates its input function, and its return value is
ignored (only called by `BaseContext.declare_function()`) - returning it
is pointless and in some sense a returned value implies not mutating the
input, so this pattern is a bit confusing.

This small refactor stops the return of the input function in order to
tidy up this ambiguity.