Remove use of `compile_isolated` from generator tests.
As title. Use of `compile_isolated` leads to hard to track down
life-time bugs such as that in #8873. Essentially the
`CodeLibrary` containing the generator destructors was being
free'd before the generators had actually had their destructors
invoked such that the code that should have been executed was
missing. This patch replaces `compile_isolated` with `jit` which
has an easier to reason about lifetime behaviour for `CodeLibrary`
instances etc.

Fixes #8873