CUDA: Enable caching functions that use CG
Caching functions that use Cooperative Groups was blocked because the
dispatcher adds a file to link, the `cudadevrt` cudalib, to the code
library, and pickling code libraries with linking files is not
supported.

To enable caching of functions that use cooperative groups, we push the
linking of `cudadevrt` into the code library. We push it in here rather
than special-casing the check for linked libraries because a recompile
of a cached function (e.g. for a different compute capability) might
require `cudadevrt` from a different location to the one found during
the initial compile; locating the library at the cubin linking time
ensures the correct version will be used.

This is based on PR #8895 which was merged into `main`.