Fix #8903. `NumbaDeprecationWarning`s raised from `@{gu,}vectorize`.
This adds localised suppressions for `NumbaDeprecationWarning`s that
would have been raised from the code path for `@{gu,}vectorize` due to
Numba's internal use of `@jit()` with the `nopython` kwarg not present.

This is considered acceptable due to the following: Deprecation warnings
from Numba's internal use of its own API should not be visible to users.
Unfortunately there's no trivial programmatic way to work around the
internal use of `@jit`-with-no-kwargs in this case as numerous code
paths go through the same operations and user code may well be relying on
"object mode" fall-back. What this patch does it makes it so that users will
not see `NumbaDeprecationWarning`s for this internal use. However, if a
user defines a `@{gu,}vectorize` function that ends up using the fall-back
to the object mode pipeline during compilation, Numba will continue to raise
numerous loud warnings about this happening which should provide a hint as
to the problem and how to fix it.

Numerous tests have been added to check combinations of kwargs/use of
default kwargs being passed through to `@{gu,}vectorize`. These tests
make sure that there are no additional `NumbaDeprecationWarning`s
raised to do with Numba's internal use of `@jit()` in `@{gu,}vectorize`
where the `nopython` kwarg is not supplied.

Closes #8903